Text.Highlighting.Kate.Types

Description

Definitions for data structures needed by highlighting-kate.

Documentation

type Context = (String, String)

A context: pair of syntax name and context name.

type ContextStack = [Context]

A stack of contexts. (Language-specific context stacks must be maintained because of IncludeRules.)

data SyntaxState

State for syntax parser.

Constructors

SyntaxState

Fields

synStContexts :: ContextStack

Stack of contexts

synStLineNumber :: Int

Number of current line

synStPrevChar :: Char

Last character parsed

synStPrevNonspace :: Bool

True if we've parsed a nonspace

synStContinuation :: Bool

True if last thing parsed is a LineContinue

synStCaseSensitive :: Bool

Language is case-sensitive

synStKeywordCaseSensitive :: Bool

Keywords are case-sensitive

synStCaptures :: [String]

List of regex captures from last capturing match

defaultSyntaxState :: SyntaxState

type Token = (TokenType, String)

A pair consisting of a list of attributes and some text.

data TokenType

Constructors

KeywordTok

DataTypeTok

DecValTok

BaseNTok

FloatTok

CharTok

StringTok

CommentTok

OtherTok

AlertTok

FunctionTok

RegionMarkerTok

ErrorTok

NormalTok

type SourceLine = [Token]

A line of source, list of labeled source items.

type KateParser = GenParser Char SyntaxState

data TokenStyle

Constructors

TokenStyle

Fields

tokenColor :: Maybe Color

tokenBackground :: Maybe Color

tokenBold :: Bool

tokenItalic :: Bool

tokenUnderline :: Bool

defStyle :: TokenStyle

data Color

Constructors

RGB Word8 Word8 Word8

class ToColor a where

Methods

toColor :: a -> Maybe Color

class FromColor a where

Methods

fromColor :: Color -> a

data Style

Constructors

Style

Fields

tokenStyles :: [(TokenType, TokenStyle)]

defaultColor :: Maybe Color

backgroundColor :: Maybe Color

lineNumberColor :: Maybe Color

lineNumberBackgroundColor :: Maybe Color

data FormatOptions

Options for formatting source code.

Constructors

FormatOptions

Fields

numberLines :: Bool

Number lines

startNumber :: Int

Number of first line

lineAnchors :: Bool

Anchors on each line number

titleAttributes :: Bool

Html titles with token types

codeClasses :: [String]

Additional classes for Html code tag

containerClasses :: [String]

Additional classes for Html container tag (pre or table depending on numberLines)

defaultFormatOpts :: FormatOptions